Compute simplex from normal vector instead of using the one from Qhull
The Qhull generated simplices are sometimes not fully dimensional in
which case they were simply skipped previously. This didn't seem to have
any effect on the completeness of the obtained description, but better
be safe. By computing the subface space from its normal vector we will
never need to skip any of the faces.