Restrict support to convex cones exclusively
Actually I wanted to finally implement rigorous support for polytopes
today – but then it felt just too painful to write all that inelegant
code.

I finally realized that it is much easier to only support convex cones
internally. However, the program can still handle polytopes in a way (or
maybe I should rather say it can now finally handle polytopes). Just
imagine one of your variables to be a constant '1' (or whatever other
value pleases you most) and don't bother that it gets treated as a
variable internally.

I guess that this might be faster even for real polytopes than what I
had in mind when I wanted to implement polytope support. This might
sound counter-intuitive since the dimension of the cone is by one higher
than that of the corresponding polytope. On the other hand, note that
for cones can be performed in the (1,1,1,…) plane effectively mitigating
this effect. Furthermore, since points in cones can be scaled we can
always have exact integer components for points.